[Web App Kafka] Fix request.timeout.ms excepted to be a 32-bit integer #13

+++ service/src/main/java/com/github/sibdevtools/service/kafka/client/api/dto/BootstrapGroupDto.java
@@ -18,6 +18,6 @@ import java.util.List;
 public class BootstrapGroupDto {
     private String code;
     private String name;
-    private Long maxTimeout;
+    private Integer maxTimeout;
     private List<String> bootstrapServers;
 }
